rename Value to Type
