+++ src/fake-schema/field/Field.md
@@ -6,27 +6,27 @@ Infrastructure to resolve a field to a fake or mock value
 
 ### resolver
 
-Returns a `resolve` function that can resolve the value of a schema type field
+Returns a `resolver` function that can resolve the value of a schema type field
 
 ### getResolver
 
-Returns an appropriate resolver function for a given field type
+Returns an appropriate `resolver` function for a given field type
 
 ### abstractTypeResolver
 
-Returns a resolver function for an abstract type
+Returns a `resolver` function for an abstract type
 
 ### fieldResolver
 
-Returns a resolver for a field, either a `Leaf` or `Complex` type resolver
+Returns a `resolver` function for a field, either a `Leaf` or `Complex` type resolver
 
 ### resolveComplexType
 
-Returns a resolver for a complex field (such as an array/list)
+Returns a `resolver` function for a complex field (such as an array/list)
 
 ### getDefaultComplexResolver
 
-Returns a generic resolver for any value not handled by specific resolver. By default this resolver function returns an empty object `{}`.
+Returns a generic resolver for any value not handled by specific `resolver`. By default this `resolver` function returns an empty object `{}`.
 
 ### getExamplesResolver
 
@@ -34,16 +34,30 @@ Returns a resolver function that resolves to a value using the `@examples` direc
 
 ### resolveLeafType
 
-Creates an instance of `PrimitiveValue` and returns a resolver function that can resolve a primitive (leaf) field to a value.
+Creates an instance of `PrimitiveValue` and returns a `resolver` function that can resolve a primitive (leaf) field to a value.
 
 ### arrayResolver
 
-Creates an instance of `ArrayValue` and returns a resolver function that can resolve an array field to a value.
+Creates an instance of `ArrayValue` and returns a `resolver` function that can resolve an array field to a value.
 
-## value
+## type
 
-See [Value](./value/Value.md)
+See [Type resolvers](./Type/Type.md)
+
+- `ArrayType`
+- `PrimitiveType`
+- `DefaultType`
 
 ## directives
 
 See [Directives](./directives/Directives.md)
+
+- `@sample`
+- `@mock`
+- `@fake`
+
+Note: The `@example` directive is so simple to resolve that it doesn't need it's own resolver class.
+
+## resolve
+
+See [Resolve](./resolve/Resolve.md)
