+++ src/fake-schema/field/type/index.ts
@@ -1,0 +1,3 @@
+export { DefaultType } from "./DefaultType";
+export { PrimitiveType } from "./PrimitiveType";
+export { ArrayType } from "./ArrayType";
