--- src/fake-schema/field/value/index.ts
@@ -1,3 +1,0 @@
-export { DefaultValue } from "./DefaultValue";
-export { PrimitiveValue } from "./PrimitiveValue";
-export { ArrayValue } from "./ArrayValue";
